a_zip_file_segments/rick_system_zip/openalgo-openalgo-theme/LangGraph-Agentic-AI-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.8
pydantic==2.5.0
pydantic-settings==2.0.3
python-multipart==0.0.6
httpx==0.25.2
pandas==2.1.3
numpy==1.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
selenium
asyncio-mqtt==0.13.0
schedule==1.2.0
loguru==0.7.2
langgraph==0.0.34  # or latest known stable
asyncpg
redis>=5.0.0
dataclasses-json
python-dotenv
